backend/requirements.txt
# Smart Livestock / Poultry Farm Management System - Backend Dependencies
# Run: pip install -r requirements.txt

flask==3.1.0
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
